Remove schemes in favour of user-requested normalization via context menu
