--- src/tuataraTMSim/NamingScheme.java
@@ -1,44 +1,0 @@
-//  ------------------------------------------------------------------
-//
-//  Copyright (c) 2006-2007 James Foulds and the University of Waikato
-//
-//  ------------------------------------------------------------------
-//  This file is part of Tuatara Turing Machine Simulator.
-//
-//  Tuatara Turing Machine Simulator is free software: you can redistribute
-//  it and/or modify it under the terms of the GNU General Public License as
-//  published by the Free Software Foundation, either version 3 of the License,
-//  or (at your option) any later version.
-//
-//  Tuatara Turing Machine Simulator is distributed in the hope that it will be
-//  useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with Tuatara Turing Machine Simulator.  If not, see
-//  <http://www.gnu.org/licenses/>.
-//
-//  author email: jf47 (at) waikato (dot) ac (dot) nz
-//
-//  ------------------------------------------------------------------
-
-package tuataraTMSim;
-
-/**
- * An enum controlling how states should be named.
- */
-public enum NamingScheme
-{
-    /**
-     * Indicates that new state names should be given general names. The choice in name is up to the
-     * creating object.
-     */
-    GENERAL,
-
-    /**
-     * Indicates that new state names should be given normalized names. This means the name of the
-     * n'th state in the machine shuold be "n" (given 0-based indexing for n).
-     */
-    NORMALIZED
-}
